# fix(sqlx-cli): do not clean sqlx during prepare