# Description

Fix `cargo sqlx prepare --workspace` running `cargo clean -p sqlx` unnecessarily. 

See https://github.com/launchbadge/sqlx/issues/2785

This should be a non-breaking change. It shouldn't affect in-workspace dependencies that happen to be named `sqlx` either.

# How to test or review this PR

- Checkout this branch and install `sqlx-cli`.

```sh
cargo install -f --path ./sqlx-cli
```

- Run `cargo sqlx prepare --workspace` in a project with crates that use `query!` macros.  
 
The timing difference is quite significant in a larger workspace, saves ~40 seconds on an incremental build.

**BEFORE:**

```bash
$ time cargo sqlx prepare --database-url "postgresql://postgres:password@localhost/example" --workspace -- --all-targets --all-features
    Cleaning [=============================] 1/1: sqlx, 0 files/folders cleaned 
...
    Finished dev [unoptimized + debuginfo] target(s) in 19.83s
query data written to .sqlx in the workspace root; please check this into version control
make sqlx-prepare  80.33s user 28.84s system 138% cpu 1:18.64 total
```

**AFTER** (this PR, note there is no cleaning step):

```bash
$ time cargo sqlx prepare --database-url "postgresql://postgres:password@localhost/example" --workspace -- --all-targets --all-features
...
    Finished dev [unoptimized + debuginfo] target(s) in 15.88s
query data written to .sqlx in the workspace root; please check this into version control
make sqlx-prepare  36.71s user 13.13s system 260% cpu 19.155 total
```