@cycraig 2023-09-27
The [unit test](https://github.com/launchbadge/sqlx/blob/051346227298b902cc21333d4bd0bc29f81222d7/sqlx-cli/src/prepare.rs#L370) is failing because it checks that the `sqlx` dependency _is_ marked to be cleaned. This seems unintentional, as the original PR for that test mentioned `sqlx` was intended to be ignored: https://github.com/launchbadge/sqlx/pull/1802#issuecomment-1099801198