@abonander 2023-11-21
Ignoring non-workspace crates seems fine to me. Those should be prepared separately.