@abonander 2024-06-21
@cycraig sorry for such a long delay on getting back to this, I tend to forget about PRs when they get pushed to the second page. Out of sight, out of mind and all that.

What do you think of this:

* `cargo sqlx prepare`: clean all workspace crates that depend on `sqlx`
* `cargo sqlx prepare --all`: clean all crates in graph that depend on `sqlx`

One thought is we could parse dep-info to know when a crate uses the macros. If you call `env!()` or `option_env!()` in a crate, it records in the dep-info that you have a dependency on that environment variable. We wouldn't even need to be able to parse the dep-info format in case it changes; we could just search for the environment variable name if it's unique enough. We could start with the crates that depend on `sqlx` and use this to filter.

It would be possible to trigger recompiles with this directly, e.g. by emitting something like `option_env!("SQLX_RECOMPILE_TRIGGER")` and setting that to a unique string every time `cargo sqlx prepare` is run. I remember trying that before but at some point I abandoned that approach, I think because it triggered a rebuild when the compiler was run without it and that got annoying. I also vaguely recall that it wasn't all that reliable.