@abonander 2024-07-16
At the end of the day, my concern is that `cargo sqlx prepare` should do what most people are going to expect it to do.

From a Cargo perspective, I suppose `cargo sqlx prepare` should be expected to prepare only the current crate, and `cargo sqlx prepare --workspace` obviously would do workspace crates. In this way, the mental model could be that it casts the smallest net possible, and then there are flags to widen that net.

But that has a couple problems:

* The constant friction from having to remember to specify `--workspace` in more complex projects.
* The lack of a clear winner for a flag to say "prepare all crates in the graph that might need it". 
    * You make a good point that `--all` is a little too ambiguous, but 5 alternatives and I'm not enamored with any of them except maybe `--all-deps`.

Instead, we could approach it from the other direction: `cargo sqlx prepare` casts the broadest net _necessary_ by default, and then we have flags to narrow that scope:

* `cargo sqlx prepare` re-runs the macros for all crates in the graph that last ran them in online mode (I have an idea of how to select for this; see below)
* `cargo sqlx prepare --workspace` restricts to just workspace crates
    * For most use-cases, this case and the previous case are equivalent.
* `cargo sqlx prepare -p <pkgspec...>`  to prepare specific crates.

While thinking about this, it occurred to me, "well duh, why don't we use files?" We can create them in the macros if they don't exist and use `include_bytes!()` in the output (or `proc_macro::tracked_path::path()` if/when it's stable) to add them to the watch list.

We can get all the selectivity we really need with three `include_bytes!()` per macro invocation:

* `<target directory>/sqlx/recompile-all`
    * Emitted by all macro invocations
* `<target directory>/sqlx/recompile-workspace`
    * Emitted by macro invocations from workspace crates only
* `$OUT_DIR/sqlx-recompile`
    * Unique path for every crate, but the file will only exist for ones that touch macros.

The contents of these files is irrelevant; I'd probably make it a timestamp. In release mode, they should be optimized out anyway.

What `cargo sqlx prepare` would then do is update the appropriate file(s) (change its contents and update its `mtime`) for a given selection:

* `cargo sqlx prepare` would touch `<target>/sqlx/recompile-all`;
* `cargo sqlx prepare --workspace` would touch `<target>/sqlx/recompile-workspace`;
* `cargo sqlx prepare -p <pkgspec>...` would touch `$OUT_DIR/sqlx-recompile` for each matched package;

then run `cargo check` for the same specification. This should pretty reliably trigger recompilation, and only for the relevant crates. This would also keep incremental compilation artifacts which would otherwise be discarded with the cleaning approach.

What do you think?