@cycraig 2024-07-31
> The constant friction from having to remember to specify `--workspace` in more complex projects.

The `--workspace` flag is usually added to a Makefile/CICD pipeline once and forgotten about, so I don't think it's a significant pain point. Such projects typically need `-- --all-targets --all-features` too, which would not change with the proposed solution.

> why don't we use files?

I'm hesitant to encourage another filesystem-based solution given the friction some developers experience with non-standard directories, e.g. https://github.com/launchbadge/sqlx/issues/1706, https://github.com/launchbadge/sqlx/issues/2395. Just saying there will likely be some teething issues.

> We can create them in the macros if they don't exist and use `include_bytes!()` in the output (or `proc_macro::tracked_path::path()` if/when it's stable) to add them to the watch list. [...] The contents of these files is irrelevant; I'd probably make it a timestamp. In release mode, they should be optimized out anyway.

It would be nice to benchmark the impact of `include_bytes!()`, if that's the only implementation option currently. The content may be negligible, but it would be nice to confirm the file handling and including bytes for hundreds to thousands of queries doesn't negatively impact developer experience.

It also sounds a bit like `invoked.timestamp`. Not sure if there's any extra insight to be gained from looking at how fingerprints work.

https://doc.rust-lang.org/nightly/nightly-rustc/cargo/core/compiler/fingerprint/index.html#fingerprint-files
>> An invoked.timestamp file whose filesystem mtime is updated every time the Unit is built. This is used for capturing the time when the build starts, to detect if files are changed in the middle of the build. See below for more details.

---

> This should pretty reliably trigger recompilation, and only for the relevant crates. This would also keep incremental compilation artifacts which would otherwise be discarded with the cleaning approach.

Thoughts: avoiding cleaning dependencies was the initial issue in this PR, and doing that with no manual configuration is certainly attractive, assuming the approach works reliably. I'm just a bit risk averse to potentially introducing a "cure worse than the disease" for a speedup of `cargo sqlx prepare`.

Excluding `sqlx` from the clean step in `sqlx prepare`, as done in [this PR](https://github.com/launchbadge/sqlx/pull/2786/commits/4509985f2bc15f26f409ef0c9f9d8c236bed7860), is an easy 1-line change that would benefit all sqlx users, with no drawbacks that I know of. That commit could be merged independently (removing the --all flag commit) while the more robust solution to avoid unnecessarily cleaning other dependencies that use sqlx (`pgvector` etc.) is implemented. Unfortunately I'm not currently in a position to contribute.